+++ changelog.md
@@ -1,2 +1,11 @@
+# 1.0.1
+- Fix Garage button position
+- Fix MoreOptionsLayer
+- Fix Stats Position in Garage
+- Add EventsPush exclusion
+- Add RecordLayer exclusion
+- Add LoadMacroLayer exclusion
+- Single Hook Approach with Alpha's Geode Utils
+
 # 1.0.0
 - Initial Release